requirements.txt
django==1.11.4
mysqlclient==1.3.10
mysql-connector==2.1.4